requirements.txt
Flask==2.1.2
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-Session==0.4.0
python-dotenv==0.20.0
Flask-WTF==1.0.1
passlib==1.7.4
PyMySQL==1.0.2
mysql-connector-python==8.0.30